test: consolidate duplicate ResourceExhaustedError spec
Section 7 — fold the unique assertions from the mislabelled src/llm/llm-integration.spec.ts (the 429/fetch-error originalError scenario, the array-filter pattern, and the try-catch capture) into src/llm/resource-exhausted.error.spec.ts, then delete the redundant integration spec. Coverage of ResourceExhaustedError is preserved (now 5 tests) and the full unit suite remains green.
